# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "댓글 및 사진 댓글 CRUD REST API. 인증은 쿠키 기반 세션으로 처리됩니다.",
        "title": "Comment API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/photo-comments": {
            "get": {
                "description": "photoSeq에 해당하는 사진 댓글 목록을 반환합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "사진 ID",
                        "name": "photoSeq",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/photocomment.PhotoCommentResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CommentSession": []
                    }
                ],
                "description": "사진에 새 댓글을 작성합니다. COMMENT_SESSION 쿠키가 필요합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 작성",
                "parameters": [
                    {
                        "description": "사진 댓글 작성 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photocomment.CreatePhotoCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/photo-comments/{commentId}": {
            "put": {
                "security": [
                    {
                        "CommentSession": []
                    },
                    {
                        "LifelogSession": []
                    }
                ],
                "description": "사진 댓글 내용을 수정합니다. 작성자(CommentSession) 또는 관리자(LifelogSession)만 수정 가능합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 수정",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "사진 댓글 수정 요청",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/photocomment.UpdatePhotoCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CommentSession": []
                    },
                    {
                        "LifelogSession": []
                    }
                ],
                "description": "사진 댓글을 소프트 삭제합니다. 작성자(CommentSession) 또는 관리자(LifelogSession)만 삭제 가능합니다.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photo-comments"
                ],
                "summary": "사진 댓글 삭제",
                "parameters": [
                    {
                        "type": "string",
                        "description": "댓글 ID (ObjectID hex)",
                        "name": "commentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "photocomment.CreatePhotoCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "photoSeq": {
                    "type": "integer"
                }
            }
        },
        "photocomment.PhotoCommentResponse": {
            "type": "object",
            "properties": {
                "authorAvatarUrl": {
                    "type": "string"
                },
                "authorEmail": {
                    "type": "string"
                },
                "authorName": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "photoSeq": {
                    "type": "integer"
                }
            }
        },
        "photocomment.UpdatePhotoCommentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "CommentSession": {
            "description": "GitHub SSO로 발급된 일반 사용자 세션 쿠키",
            "type": "apiKey",
            "name": "COMMENT_SESSION",
            "in": "cookie"
        },
        "LifelogSession": {
            "description": "lifelog(Java)에서 발급된 관리자 세션 쿠키 (읽기 전용)",
            "type": "apiKey",
            "name": "LIFELOG_SESSION",
            "in": "cookie"
        }
    }
}